- Move Game Boy classes to their own namespace. - Move SMS/GG classes to their own namespace. - Create base classes for serialisation.
