+++ wwwroot/modules/serialisers/tileSetBinarySerialiser.js
@@ -1,6 +1,4 @@
-import TileSetFactory from '../factory/tileSetFactory.js';
 import TileSet from '../models/tileSet.js';
-import TileBinarySerialiser from './tileBinarySerialiser.js';
 
 export default class TileSetBinarySerialiser {
 
@@ -11,14 +9,7 @@ export default class TileSetBinarySerialiser {
      * @returns {Uint8ClampedArray}
      */
      static serialise(tileSet) {
-        const byteLength = tileSet.length * 32;
-        const result = new Uint8ClampedArray(byteLength);
-        tileSet.getTiles().forEach((tile, index) => {
-            const resultArrayOffset = index * 32;
-            const tileAsBinary = TileBinarySerialiser.serialise(tile);
-            result.set(tileAsBinary, resultArrayOffset);
-        });
-        return result;
+        throw new Error('This method is not implemented.');
     }
 
 
@@ -28,13 +19,7 @@ export default class TileSetBinarySerialiser {
      * @returns {TileSet}
      */
      static deserialise(array) {
-        const tileSet = TileSetFactory.create();
-        for (let i = 0; i < array.length; i += 32) {
-            const arraySlice = array.slice(i, i + 32);
-            const tile = TileBinarySerialiser.deserialise(arraySlice);
-            tileSet.addTile(tile);
-        }
-        return tileSet;
+        throw new Error('This method is not implemented.');
     }
